+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.2.0] - 2022-11-11
 ### Changed
 * Added dependency on `psr/log`. As this package forms a PHP Standard
@@ -74,6 +83,7 @@ Stable release.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.2.0...HEAD
 [1.2.0]: https://github.com/DeepLcom/deepl-php/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/DeepLcom/deepl-php/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/DeepLcom/deepl-php/compare/v0.4.1...v1.0.0