+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* New languages available: Korean (`'ko'`) and Norwegian (bokmål) (`'nb'`). Add language code constants and tests.
+
+  Note: older library versions also support the new languages, this update only adds new code constants.
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.2.1] - 2023-01-25
 ### Fixed
 * Also send options in API requests even if they are default values.
@@ -79,6 +91,7 @@ Stable release.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.2.1...HEAD
 [1.2.1]: https://github.com/DeepLcom/deepl-php/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/DeepLcom/deepl-php/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/DeepLcom/deepl-php/compare/v1.0.0...v1.1.0