+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+### Added
+* Script to check our source code for license headers and a step for them in the CI.
+
 ## [1.3.0] - 2023-01-26
 ### Added
 * New languages available: Korean (`'ko'`) and Norwegian (bokmål) (`'nb'`). Add language code constants and tests.
@@ -86,6 +90,7 @@ Stable release.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.3.0...HEAD
 [1.3.0]: https://github.com/DeepLcom/deepl-php/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/DeepLcom/deepl-php/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/DeepLcom/deepl-php/compare/v1.1.0...v1.2.0