+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [Unreleased]
 ### Added
 * Script to check our source code for license headers and a step for them in the CI.
 ### Fixed
 * Fix getUsage request to be a HTTP GET request, not POST.
+* Changed document translation to poll the server every 5 seconds. This should greatly reduce observed document translation processing time.
+
 
 ## [1.3.0] - 2023-01-26
 ### Added