+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [1.4.0] - 2023-05-24
 ### Added
 * Script to check our source code for license headers and a step for them in the CI.
 * Added platform and node version information to the user-agent string that is sent with API calls, along with an opt-out.
@@ -13,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fix getUsage request to be a HTTP GET request, not POST.
 * Changed document translation to poll the server every 5 seconds. This should greatly reduce observed document translation processing time.
-
 
 
 ## [1.3.0] - 2023-01-26
@@ -97,7 +99,8 @@ Stable release.
 Initial version.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/DeepLcom/deepl-php/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/DeepLcom/deepl-php/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/DeepLcom/deepl-php/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/DeepLcom/deepl-php/compare/v1.1.0...v1.2.0